refactored indexing part
